+++ src/pato/crud/collections.py
@@ -1,4 +1,3 @@
-from fastapi import HTTPException
 from sqlalchemy import and_, case
 
 from ..models.response import FullMovie, ProcessingJobOut
@@ -13,17 +12,10 @@ from ..models.table import (
 from ..utils.database import Paged, db, paginate
 
 
-def get_tomogram(collectionId: int):
-    data = (
-        db.session.query(Tomogram)
-        .filter(Tomogram.dataCollectionId == collectionId)
-        .first()
-    )
+def get_tomograms(limit: int, page: int, collectionId: int):
+    query = db.session.query(Tomogram).filter(Tomogram.dataCollectionId == collectionId)
 
-    if data is None:
-        raise HTTPException(status_code=404, detail="Tomogram not found")
-
-    return data
+    return paginate(query, limit, page)
 
 
 def get_motion_correction(limit: int, page: int, collectionId: int) -> Paged[FullMovie]:
@@ -68,7 +60,9 @@ def get_processing_jobs(
         )
         .select_from(ProcessingJob)
         .join(AutoProcProgram)
-    ).filter(ProcessingJob.dataCollectionId == collectionId)
+        .filter(ProcessingJob.dataCollectionId == collectionId)
+        .order_by(ProcessingJob.processingJobId.desc())
+    )
 
     return paginate(
         query,