+++ tests/tomograms/test_central_slice.py
@@ -5,7 +5,7 @@ from ..conftest import mock_send
 
 def test_get_em_admin(mock_permissions, client):
     """Get central slice for tomogram"""
-    with patch("ebic.routes.tomograms.FileResponse.__call__", new=mock_send):
+    with patch("pato.routes.tomograms.FileResponse.__call__", new=mock_send):
         resp = client.get("/tomograms/1/centralSlice")
     assert resp.status_code == 200