+++ tests/tomograms/test_tomogram.py
@@ -1,4 +1,0 @@
-def test_get(mock_permissions, client):
-    """Get tomogram in a data collection belonging to user"""
-    resp = client.get("/dataCollections/6017408/tomogram")
-    assert resp.status_code == 200